whatsapp_bot/requirements.txt
twilio==8.10.0
apscheduler==3.10.4
python-dotenv==1.0.0 
flask
python-dotenv
gunicorn